+++ org/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from api.models import Member, Team, Role
+from org.models import Member, Team, Role
 from accounts.models import User
